artifacts/highrise-bot/requirements.txt
highrise-bot-sdk>=0.9.0
aiohttp>=3.8.0
